# tests/contract/openapi_snapshot.json
{
  "components": {
    "schemas": {
...
      "ActivityEvidence": {
        "additionalProperties": false,
        "properties": {
          "activity_ids": {
            "items": {
              "type": "string"
            },
            "maxItems": 200,
            "minItems": 1,
            "title": "Activity Ids",
            "type": "array"
          },
          "activity_type": {
            "enum": [
              "forum_post",
              "forum_reply",
              "wrong_question_upload",
              "exercise_attempt",
              "review_result",
              "page_view",
              "bookmark",
              "check_in"
            ],
            "title": "Activity Type",
            "type": "string"
          },
          "aggregated_count": {
            "default": 1,
            "maximum": 10000.0,
            "minimum": 1.0,
            "title": "Aggregated Count",
            "type": "integer"
          },
          "content_ref": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Ref"
          },
          "graph_node_hints": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "title": "Graph Node Hints",
            "type": "array"
          },
          "kind": {
            "const": "activity_evidence",
            "default": "activity_evidence",
            "title": "Kind",
            "type": "string"
          },
          "topic_hints": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "title": "Topic Hints",
            "type": "array"
          },
          "window_ended_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Window Ended At"
          },
          "window_started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Window Started At"
          }
        },
        "required": [
          "activity_type",
          "activity_ids"
        ],
        "title": "ActivityEvidence",
        "type": "object"
      },
...
      "CorrectMemoryCommand": {
        "additionalProperties": false,
        "properties": {
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "kind": {
            "const": "correct_memory",
            "default": "correct_memory",
            "title": "Kind",
            "type": "string"
          },
          "memory_id": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Memory Id",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "replacement": {
            "discriminator": {
              "mapping": {
                "learner": "#/components/schemas/LearnerReplacement",
                "mastery": "#/components/schemas/MasteryReplacement"
              },
              "propertyName": "replacement_type"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/LearnerReplacement"
              },
              {
                "$ref": "#/components/schemas/MasteryReplacement"
              }
            ],
            "title": "Replacement"
          }
        },
        "required": [
          "memory_id",
          "expected_version",
          "replacement"
        ],
        "title": "CorrectMemoryCommand",
        "type": "object"
      },
      "CursorPage_DeletedMemoryItem_": {
        "properties": {
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/DeletedMemoryItem"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor",
          "has_more"
        ],
        "title": "CursorPage[DeletedMemoryItem]",
        "type": "object"
      },
...
      "ForgetMemoryCommand": {
        "additionalProperties": false,
        "properties": {
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "kind": {
            "const": "forget_memory",
            "default": "forget_memory",
            "title": "Kind",
            "type": "string"
          },
          "memory_id": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Memory Id",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "required": [
          "memory_id",
          "expected_version"
        ],
        "title": "ForgetMemoryCommand",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MemoryOperationResult": {
        "properties": {
          "cancelled_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancelled At"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PublicError"
              },
              {
                "type": "null"
              }
            ]
          },
          "graph_state_changes": {
            "items": {
              "$ref": "#/components/schemas/GraphStateChangeView"
            },
            "title": "Graph State Changes",
            "type": "array"
          },
          "mutations": {
            "items": {
              "$ref": "#/components/schemas/MutationResult"
            },
            "title": "Mutations",
            "type": "array"
          },
          "operation_id": {
            "format": "uuid",
            "title": "Operation Id",
            "type": "string"
          },
          "operation_type": {
            "enum": [
              "conversation_evidence",
              "activity_evidence",
              "correct_memory",
              "forget_memory",
              "restore_memory",
              "override_learner_profile",
              "review_candidate",
              "set_graph_state",
              "project_summary_to_graph",
              "rebuild_index",
              "verify_checksums",
              "purge_tombstones",
              "cleanup_orphan_versions",
              "cleanup_checkpoints",
              "purge_account_memory"
            ],
            "title": "Operation Type",
            "type": "string"
          },
          "review_candidate_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Review Candidate Ids",
            "type": "array"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "retry_wait",
              "succeeded",
              "needs_review",
              "dead_letter",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "operation_id",
          "status",
          "operation_type",
          "created_at",
          "updated_at"
        ],
        "title": "MemoryOperationResult",
        "type": "object"
      },
...
      "RestoreMemoryCommand": {
        "additionalProperties": false,
        "properties": {
          "deleted_version": {
            "minimum": 1.0,
            "title": "Deleted Version",
            "type": "integer"
          },
          "kind": {
            "const": "restore_memory",
            "default": "restore_memory",
            "title": "Kind",
            "type": "string"
          },
          "memory_id": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Memory Id",
            "type": "string"
          }
        },
        "required": [
          "memory_id",
          "deleted_version"
        ],
        "title": "RestoreMemoryCommand",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "MemoryManagerGraph API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/memory/commands/correct": {
      "post": {
        "operationId": "correct_memory_api_v1_memory_commands_correct_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CorrectMemoryCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryOperationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Correct Memory",
        "tags": [
          "memory"
        ]
      }
    },
    "/api/v1/memory/commands/forget": {
      "post": {
        "operationId": "forget_memory_api_v1_memory_commands_forget_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgetMemoryCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryOperationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Forget Memory",
        "tags": [
          "memory"
        ]
      }
    },
    "/api/v1/memory/commands/restore": {
      "post": {
        "operationId": "restore_memory_api_v1_memory_commands_restore_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestoreMemoryCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryOperationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Restore Memory",
        "tags": [
          "memory"
        ]
      }
    },
    "/api/v1/memory/deleted": {
      "get": {
        "description": "30 天恢复窗口内的已删除记忆（§19.4）；cursor 绑定路由/用户/筛选。",
        "operationId": "list_deleted_api_v1_memory_deleted_get",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 1000,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CursorPage_DeletedMemoryItem_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Deleted",
        "tags": [
          "memory"
        ]
      }
    },
    "/api/v1/memory/events": {
      "post": {
        "description": "学习证据入队（§19.1）；P2/P3 不走快速路径，固定 202。",
        "operationId": "submit_event_api_v1_memory_events_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "discriminator": {
                  "mapping": {
                    "activity_evidence": "#/components/schemas/ActivityEvidence",
                    "conversation_evidence": "#/components/schemas/ConversationEvidence"
                  },
                  "propertyName": "kind"
                },
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/ConversationEvidence"
                  },
                  {
                    "$ref": "#/components/schemas/ActivityEvidence"
                  }
                ],
                "title": "Request"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryOperationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Event",
        "tags": [
          "memory"
        ]
      }
    },
...
    "/api/v1/memory/learner": {
      "get": {
        "operationId": "get_learner_api_v1_memory_learner_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearnerMemoryView"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Learner",
        "tags": [
          "memory"
        ]
      },
      "put": {
        "operationId": "override_learner_api_v1_memory_learner_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OverrideLearnerProfileCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryOperationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Override Learner",
        "tags": [
          "memory"
        ]
      }
    },
    "/api/v1/memory/mastery/{topic_key}": {
      "get": {
        "operationId": "get_mastery_api_v1_memory_mastery__topic_key__get",
        "parameters": [
          {
            "in": "path",
            "name": "topic_key",
            "required": true,
            "schema": {
              "title": "Topic Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MasteryMemoryView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Mastery",
        "tags": [
          "memory"
        ]
      }
    },
    "/api/v1/memory/memories/{memory_id}": {
      "get": {
        "operationId": "get_memory_by_id_api_v1_memory_memories__memory_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "memory_id",
            "required": true,
            "schema": {
              "title": "Memory Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/LearnerMemoryView"
                    },
                    {
                      "$ref": "#/components/schemas/MasteryMemoryView"
                    }
                  ],
                  "title": "Response Get Memory By Id Api V1 Memory Memories  Memory Id  Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Memory By Id",
        "tags": [
          "memory"
        ]
      }
    },
...
  }
}